Update parsePayoff to support float values in payoff strings
The parsePayoff function now accepts float values in payoff expressions like "1|4x0.25|19x0".
Updated Market.ParsedPayoff field type from []int to []float64 to support decimal payoff values.

+++ pkg/outrights/types.go
@@ -30,7 +30,7 @@ type Event struct {
 type Market struct {
 	Name         string    `json:"name"`
 	Payoff       string    `json:"payoff"`
-	ParsedPayoff []int     `json:"-"` // Parsed version, not serialized
+	ParsedPayoff []float64 `json:"-"` // Parsed version, not serialized
 	Teams        []string  `json:"teams,omitempty"`
 	Include      []string  `json:"include,omitempty"`
 	Exclude      []string  `json:"exclude,omitempty"`
